chore(report): change routes and final adjusts
