+++ src/views/coupon-report.html
@@ -1,0 +1,86 @@
+<div class="main-content">
+  <div class="card">
+    <div class="card-title">
+      <big class="text-secondary mr-1">
+        <i class="icon fa fa-percent"></i>
+      </big>
+      <big class="i18n fw-400">
+        <span data-lang="en_us">Coupon report</span>
+        <span data-lang="pt_br">Relatório de Cupons</span>
+      </big>
+
+      <div class="input-group float-right" id="datepicker-coupon" data-provide="datepicker" style="width: 300px">
+        <div class="input-group-prepend">
+          <span class="input-group-text i18n">
+            <span data-lang="en_us">From</span>
+            <span data-lang="pt_br">De</span>
+          </span>
+        </div>
+        <input type="text" class="form-control" data-when="start">
+        <div class="input-group-prepend input-group-append">
+          <span class="input-group-text i18n">
+            <span data-lang="en_us">To</span>
+            <span data-lang="pt_br">Até</span>
+          </span>
+        </div>
+        <input type="text" class="form-control" data-when="end">
+      </div>
+    </div>
+
+    <div class="card-body">
+      <div class="row">
+        <div class="col-12">
+          <div class="card">
+            <div class="card-content">
+              <div class="card-body">
+                <div class="row">
+                  <div class="col-md-12">
+                    <div class="mb-3 mt-4">
+                      <a href="#" class="btn btn-info" id="export-coupon">
+                        <span class="i18n">
+                          <span data-lang="pt_br">Exportar</span>
+                          <span data-lang="en_us">Export</span>
+                        </span>
+                      </a>
+                    </div>
+                    <table id="coupon-list" class="display" style="width:100%">
+                      <thead>
+                          <tr>
+                              <th class="i18n">
+                                <span data-lang="en_us">Coupon</span>
+                                <span data-lang="pt_br">Cupom</span>
+                              </th>
+                              <th class="i18n">
+                                <span data-lang="en_us">Quantity</span>
+                                <span data-lang="pt_br">Quantidade</span>
+                              </th>
+                              <th class="i18n">
+                                <span data-lang="en_us">Discount</span>
+                                <span data-lang="pt_br">Desconto</span>
+                              </th>
+                              <th class="i18n">
+                                <span data-lang="en_us">Freight</span>
+                                <span data-lang="pt_br">Frete</span>
+                              </th>
+                              <th class="i18n">
+                                <span data-lang="en_us">Revenue</span>
+                                <span data-lang="pt_br">Receita</span>
+                              </th>
+                          </tr>
+                        </thead>
+                      <tbody></tbody>
+                    </table>
+                    <span class="i18n">
+                      <strong data-lang="pt_br">Observação: Se sua loja tiver a possibilidade de acumular descontos de diferentes campanhas, o valor do desconto mostrado irá incluir essas outras modalidades.</strong>
+                      <strong data-lang="en_us">Note: If your store has the possibility of accumulating discounts from different campaigns, the discount amount shown will include these other modalities.</strong>
+                    </span>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
